ITA-Abdykarov-Sprint-5-changes

See merge request ilias.abdykarov/ita-2021-java!8

+++ src/main/java/cz/abdykili/eshop/service/impl/ProductServiceImpl.java
@@ -1,12 +1,16 @@
 package cz.abdykili.eshop.service.impl;
 
 import cz.abdykili.eshop.domain.Product;
+import cz.abdykili.eshop.exception.ProductNotFoundException;
 import cz.abdykili.eshop.model.ProductResponseDto;
 import cz.abdykili.eshop.model.ProductRequestDto;
 import cz.abdykili.eshop.repository.ProductRepository;
 import cz.abdykili.eshop.service.ProductService;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -20,48 +24,79 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
-
+    private final Logger logger = LoggerFactory.getLogger(ProductServiceImpl.class);
     @Autowired
     private final ProductRepository productRepository;
 
     @Override
     public List<ProductResponseDto> findAll(){
-        return productRepository.findAll().stream()
+        List<ProductResponseDto> products = productRepository.findAll().stream()
                 .map(p -> mapToResponse(p))
                 .collect(Collectors.toList());
+
+        logger.info("Returning {} products from repository", products.size());
+        logger.debug("Returning the products {}", products);
+
+        return products;
     }
 
     @Override
     public ProductResponseDto findProduct(Long id){
         final Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product with id " + id + " not found!"));
+                .orElseThrow(() -> new ProductNotFoundException("Product with id " + id + " not found!",
+                        HttpStatus.NOT_FOUND));
+        ProductResponseDto responseProduct = mapToResponse(product);
 
-        return mapToResponse(product);
+        logger.info("Returning the product with id {} from repository", product.getId());
+        logger.debug("Returning the product {}", responseProduct);
+
+        return responseProduct;
     }
 
     @Override
     public ProductResponseDto saveProduct(ProductRequestDto productRequestDto){
+        logger.info("Saving a new product");
+        logger.debug("Saving a new product - Incoming payload {} ", productRequestDto);
+
         Product newProduct = mapToEntity(productRequestDto);
         final Product savedProduct = productRepository.save(newProduct);
-        return mapToResponse(savedProduct);
+        ProductResponseDto productResponseDto = mapToResponse(savedProduct);
+
+        logger.debug("Saving a new product - Incoming payload {}, outgoing payload {} ", productRequestDto, productResponseDto);
+
+        return productResponseDto;
     }
 
     @Override
     public ProductResponseDto updateProduct(ProductRequestDto productRequestDto, Long id){
-        final Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product with id " + id + " not found!"));
-        Product productToSave = mapToEntity(productRequestDto)
-                .setId(product.getId());
-        final Product updatedProduct = productRepository.save(productToSave);
+        logger.info("Updating an existing product with id {}", id);
+        logger.debug("Updating an existing product with id {} - Incoming payload {} ", id, productRequestDto);
+        if(productRepository.existsById(id)){
 
-        return mapToResponse(updatedProduct);
+            Product productToSave = mapToEntity(productRequestDto)
+                    .setId(id);
+
+            final Product updatedProduct = productRepository.save(productToSave);
+            ProductResponseDto productResponseDto = mapToResponse(updatedProduct);
+
+            logger.debug("Updating an existing product | incoming payload {}, outgoing payload {} ", productRequestDto, productResponseDto);
+
+            return productResponseDto;
+        }
+        else{
+            logger.error("Updating an existing product with id {} | Entity with id {} was not found", id, id);
+        }
+        return null;
     }
 
     @Override
     public void deleteProduct(Long id){
-        final Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product with id " + id + " not found!"));
-        productRepository.deleteById(id);
+        logger.info("Deleting the product with id {}", id);
+        if(productRepository.existsById(id)){
+            productRepository.deleteById(id);
+        }else{
+            logger.error("Deleting the product with id {} | Entity with id {} was not found", id, id);
+        }
     }
 
     private Product mapToEntity(ProductRequestDto productRequestDto){
@@ -82,5 +117,6 @@ public class ProductServiceImpl implements ProductService {
                 .setStock(savedProduct.getStock())
                 .setName(savedProduct.getName());
     }
+
 
 }
